Move addAmination to dom
